+++ tests/header_window.cpp
@@ -4,3 +4,8 @@ class HeaderWindow final : public mwtl::Window<HeaderWindow> {
 public:
     void BuildUI() {}
 };
+
+class HeaderSimpleWindow final : public mwtl::WindowBase {
+public:
+    void BuildUI() override {}
+};